requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-Blinka==8.9.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-framebuf==1.4.14
adafruit-circuitpython-register==1.9.15
adafruit-circuitpython-requests==1.12.11
adafruit-circuitpython-shtc3==1.1.11
adafruit-circuitpython-ssd1306==2.12.12
adafruit-circuitpython-typing==1.8.3
adafruit-extended-bus==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.37.0
Adafruit-PureIO==1.1.9
LoRaRF==1.4.0
netifaces==0.11.0
Pillow==9.3.0
psutil==5.9.4
pyftdi==0.54.0
pynmeagps==1.0.16
pyrtcm==0.3.1
pyserial==3.5
pyubx2==1.2.17
pyusb==1.2.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
spidev==3.6
sysv-ipc==1.1.0
typing-extensions==4.4.0
